pulumi-python/requirements.txt
pulumi>=2.0.0,<3.0.0
pulumi-aws>=3.2.0,<4.0.0
pulumi-docker
pulumi-kubernetes
PyYAML
